+++ .run/IzumBankApplication.run.xml
@@ -1,0 +1,15 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="IzumBankApplication" type="SpringBootApplicationConfigurationType" factoryName="Spring Boot" nameIsGenerated="true">
+    <module name="izum-bank-app.main" />
+    <option name="SPRING_BOOT_MAIN_CLASS" value="org.matrix.izumbankapp.IzumBankApplication" />
+    <extension name="coverage">
+      <pattern>
+        <option name="PATTERN" value="org.matrix.izumbankapp.*" />
+        <option name="ENABLED" value="true" />
+      </pattern>
+    </extension>
+    <method v="2">
+      <option name="Make" enabled="true" />
+    </method>
+  </configuration>
+</component>
