+++ src/main/java/org/matrix/izumbankapp/service/AccountService.java
@@ -1,5 +1,7 @@
 package org.matrix.izumbankapp.service;
 
+import org.matrix.izumbankapp.enumeration.accounts.AccountStatus;
+import org.matrix.izumbankapp.enumeration.accounts.AccountType;
 import org.matrix.izumbankapp.model.accounts.*;
 import org.matrix.izumbankapp.model.auth.AccountStatusUpdate;
 import org.matrix.izumbankapp.model.auth.ResponseDto;
@@ -8,6 +10,7 @@ import org.matrix.izumbankapp.model.users.UserAccountsResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.time.LocalDate;
 import java.util.List;
 
 
@@ -37,11 +40,6 @@ public interface AccountService {
 
     UserAccountsResponse getUserByAccountNumber(String accountNumber);
 
-    ResponseDto transferToAccount(TransferMoneyRequest transferMoneyRequest);
-
-    ResponseDto withdrawal(WithdrawalRequest withdrawalRequest);
-
-    ResponseDto createDepositAccount(DepositRequest depositRequest);
 
     void saveAccount(AccountResponse account);
